broke down threadpool.cpp and threadpool.h into multiple files, each file coresponds to one class now
