+++ Global.h
@@ -1,0 +1,5 @@
+#pragma once
+
+const int DEFAULT_POOL_SIZE = 10;
+const int STARTED = 0;
+const int STOPPED = 1;
